pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-clinic_scheduler"
version = "0.1.0"
description = "Clinic Scheduler environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m clinic_scheduler.server.app
server = "clinic_scheduler.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["clinic_scheduler", "clinic_scheduler.server"]
package-dir = { "clinic_scheduler" = ".", "clinic_scheduler.server" = "server" }